stream-processor/requirements.txt
apache-flink==1.18.1
psycopg2-binary==2.9.9
pika==1.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1 